fix artefacts

+++ includes/lgs.h
@@ -35,16 +35,16 @@ t_symbol*			create_symbols_list(char const * str, size_t str_len, int32_t* left)
 t_symbol*			create_symbols_list_nospace(char const * str, size_t str_len, int32_t * left,
 												 int32_t* top, int32_t* bottom);
 
-unsigned char*		string_create_image(char const * str, int pix_size, size_t *width, size_t *height);
+unsigned char*		string_create_image(char const * str, int pix_size, int* width, int* height);
 
 unsigned char*		string_create_image_nospace(char const * str, int pix_size,
-													  size_t* width, size_t* height);
+													  int* width, int* height);
 
 char				set_vao_vbo_ebo(t_string* string);
 
 void				string_add(t_string** lst, t_string* elem);
 
-t_string*			string_create_vao(unsigned char const * image, size_t width, size_t height, GLFWwindow* window);
+t_string*			string_create_vao(unsigned char const * image, int width, int height, GLFWwindow* window);
 
 GLchar const *const	get_standart_vertix_shader(void);
